extract Bazel build file parsing utility
